modularization: extract CoverageData
the `CoverageData.summary()` method was removed and put into
`print_text_report()` as it is mostly presentational.
